schema(reindex): consolidate search_settings port columns into a single migration
